DB connection

+++ Model/Entities/Group.cs
@@ -5,8 +5,8 @@ namespace Model
 	public class Group
 	{
 		public string Name { get; set; }
-		public int Size { get; set; }
-		public int Course { get; set; }
+		public long Size { get; set; }
+		public long Course { get; set; }
 
 		public ICollection<Student> Students { get; set; }
 		public ICollection<Timetable> Lessons { get; set; }
